Speed Text
The code creates a simple Typing Speed Test using Tkinter. It generates random words from a list and 
the user types them into an Entry widget. The code tracks the user's accuracy and speed (in words per minute) 
and displays the results in a Label widget. The user can start a new test by clicking the Start button
and can reset the test by clicking the Reset button.